Rework API url configuration to use DRF-versioning
Previously, we were just having the "version" be a hard-coded URL
path prefix.  However, now we are using DRF's built-in versioning
tools so that it is aware of the actual version.  Consequently, we
also have to tell DRF-Spectacular about the version, but this lets us
get a UI with correct version information displayed.